+++ data_generator/predator.h
@@ -11,21 +11,21 @@ class prey;
 using namespace std;
 class  predator : public animal {
 public:
-predator(float startX, float startY);
+predator(double startX, double startY);
 void update();
-float getX() const { return x; };
-float getY() const { return y; };
-float getVX() const { return vx; };
-float getVY() const { return vy; };
+double getX() const { return x; };
+double getY() const { return y; };
+double getVX() const { return vx; };
+double getVY() const { return vy; };
 //matrix of distances from prey in close proximity
 std::pair<double, double> look_for_prey(const std::vector<prey>& vec_prey) const;
 //calculation of change in direction
 void persue(std::pair<double, double>& r_min, predator& self);
 
 protected: 
-float x, y;           
-float vx, vy;   
-float v_size = 15.0;   
+double x, y;           
+double vx, vy;   
+double v_size = 15.0;   
 int sight = 300;   //distance required for change in direction    
 private:
